--- modules/data/src/main/java/net/pantasystem/milktea/data/gettters/README.md
@@ -1,4 +1,0 @@
-# gettersパッケージ
-## このパッケージについて
-このパッケージは依存するデータのRepositoryクラスのsuspend関数と、  
-Event発行クラスから発行されるFlowをいい感じに合わせて関数やインターフェースやクラスが作成されます。
