+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/messaging/MessageRepositoryImpl.kt
@@ -6,7 +6,6 @@ import net.pantasystem.milktea.api.misskey.messaging.MessageAction
 import net.pantasystem.milktea.api.misskey.messaging.MessageDTO
 import net.pantasystem.milktea.common.throwIfHasError
 import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
-import net.pantasystem.milktea.data.gettters.MessageAdder
 import net.pantasystem.milktea.model.account.AccountRepository
 import net.pantasystem.milktea.model.messaging.CreateMessage
 import net.pantasystem.milktea.model.messaging.Message
