+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/streaming/StreamingMainMessageEventDispatcher.kt
@@ -1,7 +1,7 @@
 package net.pantasystem.milktea.data.infrastructure.streaming
 
 import net.pantasystem.milktea.api_streaming.ChannelBody
-import net.pantasystem.milktea.data.gettters.MessageAdder
+import net.pantasystem.milktea.data.infrastructure.messaging.MessageAdder
 import net.pantasystem.milktea.data.infrastructure.messaging.MessageDataSource
 import net.pantasystem.milktea.model.account.Account
 import javax.inject.Inject
